refactor prompts into separate modules
