+++ backend/services/prompts/__init__.py
@@ -1,0 +1,21 @@
+from .theme import create_theme_instructions
+from .research_planner import create_research_planner_instructions
+from .researcher import create_researcher_instructions
+from .research_synthesizer import create_research_synthesizer_instructions
+from .outline import create_outline_instructions
+from .section_writer import create_section_writer_instructions
+from .editor import create_editor_instructions
+from .persona_generator import create_persona_generator_instructions
+from .serp_keyword_analysis import create_serp_keyword_analysis_instructions
+
+__all__ = [
+    "create_theme_instructions",
+    "create_research_planner_instructions",
+    "create_researcher_instructions",
+    "create_research_synthesizer_instructions",
+    "create_outline_instructions",
+    "create_section_writer_instructions",
+    "create_editor_instructions",
+    "create_persona_generator_instructions",
+    "create_serp_keyword_analysis_instructions",
+]
