+++ backend/services/prompts/editor.py
@@ -1,0 +1,62 @@
+# -*- coding: utf-8 -*-
+from typing import Callable, Awaitable
+from agents import Agent, RunContextWrapper
+from services.context import ArticleContext
+
+
+def create_editor_instructions(base_prompt: str) -> Callable[[RunContextWrapper[ArticleContext], Agent[ArticleContext]], Awaitable[str]]:
+    async def dynamic_instructions_func(ctx: RunContextWrapper[ArticleContext], agent: Agent[ArticleContext]) -> str:
+        if not ctx.context.full_draft_html:
+            raise ValueError("編集対象のドラフト記事がありません。")
+        if not ctx.context.research_report:
+            raise ValueError("参照すべきリサーチレポートがありません。")
+        if not ctx.context.selected_detailed_persona:
+            raise ValueError("編集のための詳細なペルソナが選択されていません。")
+        persona_description = ctx.context.selected_detailed_persona
+
+        research_context_str = f"リサーチ要約: {ctx.context.research_report.overall_summary[:500]}...\n"
+        research_context_str += "主要なキーポイントと出典:\n"
+        for kp in ctx.context.research_report.key_points:
+            sources_str = ", ".join([f"[{url.split('/')[-1] if url.split('/')[-1] else url}]({url})" for url in kp.supporting_sources])
+            research_context_str += f"- {kp.point} (出典: {sources_str})\n"
+        research_context_str += f"参照した全情報源URL数: {len(ctx.context.research_report.all_sources)}\n"
+
+        full_prompt = f"""{base_prompt}
+
+--- 編集対象記事ドラフト (HTML) ---
+```html
+{ctx.context.full_draft_html[:15000]}
+{ "... (以下省略)" if len(ctx.context.full_draft_html) > 15000 else "" }
+```
+---
+
+--- 記事の要件 ---
+タイトル: {ctx.context.generated_outline.title if ctx.context.generated_outline else 'N/A'}
+キーワード: {', '.join(ctx.context.selected_theme.keywords) if ctx.context.selected_theme else 'N/A'}
+ターゲットペルソナ詳細:\n{persona_description}
+目標文字数: {ctx.context.target_length or '指定なし'}
+トーン: {ctx.context.generated_outline.suggested_tone if ctx.context.generated_outline else 'N/A'}
+企業スタイルガイド: {ctx.context.company_style_guide or '指定なし'}
+--- 詳細なリサーチ情報 ---
+{research_context_str[:10000]}
+{ "... (以下省略)" if len(research_context_str) > 10000 else "" }
+---
+
+**重要:**
+- 上記のドラフトHTMLをレビューし、記事の要件と**詳細なリサーチ情報**に基づいて推敲・編集してください。
+- **特に、文章全体がターゲットペルソナ「{persona_description}」にとって自然で、親しみやすく、分かりやすい言葉遣いになっているか** を重点的に確認してください。機械的な表現や硬い言い回しがあれば、より人間味のある表現に修正してください。
+- チェックポイント:
+    - 全体の流れと一貫性
+    - 各セクションの内容の質と正確性 (**リサーチ情報との整合性、事実確認**)
+    - 文法、スペル、誤字脱字
+    - 指示されたトーンとスタイルガイドの遵守 (**自然さ、親しみやすさ重視**)
+    - ターゲットペルソナへの適合性
+    - SEO最適化（キーワードの自然な使用、見出し構造）
+    - **含まれているHTMLリンク (`<a>` タグ) がリサーチ情報に基づいており、適切かつ自然に使用されているか。リンク切れや不適切なリンクがないか。**
+    - 人間らしい自然な文章表現、独創性
+    - HTML構造の妥当性
+- 必要な修正を直接HTMLに加えてください。
+- あなたの応答は必ず `RevisedArticle` 型のJSON形式で、`final_html_content` に編集後の完全なHTML文字列を入れて出力してください。
+"""
+        return full_prompt
+    return dynamic_instructions_func
