+++ backend/services/prompts/research_synthesizer.py
@@ -1,0 +1,47 @@
+# -*- coding: utf-8 -*-
+from typing import Callable, Awaitable
+from agents import Agent, RunContextWrapper
+from services.context import ArticleContext
+
+
+def create_research_synthesizer_instructions(base_prompt: str) -> Callable[[RunContextWrapper[ArticleContext], Agent[ArticleContext]], Awaitable[str]]:
+    async def dynamic_instructions_func(ctx: RunContextWrapper[ArticleContext], agent: Agent[ArticleContext]) -> str:
+        if not ctx.context.research_query_results:
+            raise ValueError("要約するためのリサーチ結果がありません。")
+
+        results_str = ""
+        all_sources_set = set()
+        for i, result in enumerate(ctx.context.research_query_results):
+            results_str += f"--- クエリ結果 {i+1} ({result.query}) ---\n"
+            results_str += f"要約: {result.summary}\n"
+            results_str += "詳細な発見:\n"
+            for finding in result.detailed_findings:
+                results_str += f"- 抜粋: {finding.snippet_text}\n"
+                results_str += f"  出典: [{finding.source_title or finding.source_url}]({finding.source_url})\n"
+                all_sources_set.add(finding.source_url)
+            results_str += "\n"
+
+        all_sources_list = sorted(list(all_sources_set))
+
+        full_prompt = f"""{base_prompt}
+
+--- リサーチ対象テーマ ---
+{ctx.context.selected_theme.title if ctx.context.selected_theme else 'N/A'}
+
+--- 収集されたリサーチ結果 (詳細) ---
+{results_str[:15000]}
+{ "... (以下省略)" if len(results_str) > 15000 else "" }
+---
+
+**重要:**
+- 上記の詳細なリサーチ結果全体を分析し、記事執筆に役立つように情報を統合・要約してください。
+- 以下の要素を含む**実用的で詳細なリサーチレポート**を作成してください:
+    - `overall_summary`: リサーチ全体から得られた主要な洞察やポイントの要約。
+    - `key_points`: 記事に含めるべき重要なポイントや事実をリスト形式で記述し、各ポイントについて**それを裏付ける情報源URL (`supporting_sources`)** を `KeyPoint` 形式で明確に紐付けてください。
+    - `interesting_angles`: 記事を面白くするための切り口や視点のアイデアのリスト形式。
+    - `all_sources`: 参照した全ての情報源URLのリスト（重複削除済み、可能であれば重要度順）。
+- レポートは論文調ではなく、記事作成者がすぐに使えるような分かりやすい言葉で記述してください。
+- あなたの応答は必ず `ResearchReport` 型のJSON形式で出力してください。
+"""
+        return full_prompt
+    return dynamic_instructions_func
