+++ backend/services/prompts/researcher.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+from typing import Callable, Awaitable
+from agents import Agent, RunContextWrapper
+from services.context import ArticleContext
+
+
+def create_researcher_instructions(base_prompt: str) -> Callable[[RunContextWrapper[ArticleContext], Agent[ArticleContext]], Awaitable[str]]:
+    async def dynamic_instructions_func(ctx: RunContextWrapper[ArticleContext], agent: Agent[ArticleContext]) -> str:
+        if not ctx.context.research_plan or ctx.context.current_research_query_index >= len(ctx.context.research_plan.queries):
+            raise ValueError("有効なリサーチプランまたは実行すべきクエリがありません。")
+
+        current_query = ctx.context.research_plan.queries[ctx.context.current_research_query_index]
+
+        full_prompt = f"""{base_prompt}
+
+--- 現在のリサーチタスク ---
+記事テーマ: {ctx.context.research_plan.topic}
+今回の検索クエリ: "{current_query.query}"
+このクエリの焦点: {current_query.focus}
+---
+
+**重要:**
+- 上記の検索クエリを使用して `web_search` ツールを実行してください。
+- 検索結果を**深く分析**し、記事テーマとクエリの焦点に関連する**具体的な情報、データ、主張、引用**などを**詳細に抽出**してください。
+- 抽出した各情報について、**最も信頼性が高く具体的な出典元URLとそのタイトル**を特定し、`SourceSnippet` 形式でリスト化してください。単なる検索結果一覧のURLではなく、情報が実際に記載されているページのURLを重視してください。公式HPや信頼できる情報源を優先してください。
+- 検索結果全体の**簡潔な要約 (summary)** も生成してください。
+- あなたの応答は必ず `ResearchQueryResult` 型のJSON形式で出力してください。他のテキストは一切含めないでください。
+- **`save_research_snippet` ツールは使用しないでください。**
+"""
+        return full_prompt
+    return dynamic_instructions_func
