+++ backend/services/prompts/section_writer.py
@@ -1,0 +1,84 @@
+# -*- coding: utf-8 -*-
+from typing import Callable, Awaitable
+from agents import Agent, RunContextWrapper
+from services.context import ArticleContext
+
+
+def create_section_writer_instructions(base_prompt: str) -> Callable[[RunContextWrapper[ArticleContext], Agent[ArticleContext]], Awaitable[str]]:
+    async def dynamic_instructions_func(ctx: RunContextWrapper[ArticleContext], agent: Agent[ArticleContext]) -> str:
+        if not ctx.context.generated_outline or ctx.context.current_section_index >= len(ctx.context.generated_outline.sections):
+            raise ValueError("有効なアウトラインまたはセクションインデックスがありません。")
+        if not ctx.context.research_report:
+            raise ValueError("参照すべきリサーチレポートがありません。")
+        if not ctx.context.selected_detailed_persona:
+            raise ValueError("セクション執筆のための詳細なペルソナが選択されていません。")
+        persona_description = ctx.context.selected_detailed_persona
+
+        target_section = ctx.context.generated_outline.sections[ctx.context.current_section_index]
+        target_index = ctx.context.current_section_index
+        target_heading = target_section.heading
+
+        section_target_chars = None
+        if ctx.context.target_length and len(ctx.context.generated_outline.sections) > 0:
+            total_sections = len(ctx.context.generated_outline.sections)
+            estimated_total_body_chars = ctx.context.target_length * 0.8
+            section_target_chars = int(estimated_total_body_chars / total_sections)
+
+        outline_context = "\n".join([f"{i+1}. {s.heading}" for i, s in enumerate(ctx.context.generated_outline.sections)])
+
+        research_context_str = f"リサーチ要約: {ctx.context.research_report.overall_summary[:500]}...\n"
+        research_context_str += "主要なキーポイントと出典:\n"
+        for kp in ctx.context.research_report.key_points:
+            sources_str = ", ".join([f"[{url.split('/')[-1] if url.split('/')[-1] else url}]({url})" for url in kp.supporting_sources])
+            research_context_str += f"- {kp.point} (出典: {sources_str})\n"
+        research_context_str += f"参照した全情報源URL数: {len(ctx.context.research_report.all_sources)}\n"
+
+        company_style_guide = ctx.context.company_style_guide or '指定なし'
+
+        full_prompt = f"""{base_prompt}
+
+--- 記事全体の情報 ---
+記事タイトル: {ctx.context.generated_outline.title}
+記事全体のキーワード: {', '.join(ctx.context.selected_theme.keywords) if ctx.context.selected_theme else 'N/A'}
+記事全体のトーン: {ctx.context.generated_outline.suggested_tone}
+ターゲットペルソナ詳細:\n{persona_description}
+企業スタイルガイド: {company_style_guide}
+記事のアウトライン（全体像）:
+{outline_context}
+--- 詳細なリサーチ情報 ---
+{research_context_str[:10000]}
+{ "... (以下省略)" if len(research_context_str) > 10000 else "" }
+---
+
+--- **あなたの現在のタスク** ---
+あなたは **セクションインデックス {target_index}**、見出し「**{target_heading}**」の内容をHTML形式で執筆するタスク**のみ**を担当します。
+このセクションの目標文字数: {section_target_chars or '指定なし（流れに合わせて適切に）'}
+---
+
+--- **【最重要】執筆スタイルとトーンについて** ---
+あなたの役割は、単に情報をHTMLにするだけでなく、**まるで経験豊富な友人が以下のペルソナ「{persona_description}」に語りかけるように**、親しみやすく、分かりやすい文章でセクションを執筆することです。
+- **日本の一般的なブログ記事やコラムのような、自然で人間味あふれる、温かいトーン**を心がけてください。堅苦しい表現や機械的な言い回しは避けてください。
+- 読者に直接語りかけるような表現（例：「〜だと思いませんか？」「まずは〜から始めてみましょう！」「〜なんてこともありますよね」）や、共感を誘うような言葉遣いを積極的に使用してください。
+- 専門用語は避け、どうしても必要な場合は簡単な言葉で補足説明を加えてください。箇条書きなども活用し、情報を整理して伝えると良いでしょう。
+- 可能であれば、具体的な体験談（想像でも構いません）や、読者が抱きそうな疑問に答えるような形で内容を構成すると、より読者の心に響きます。
+- 企業スタイルガイド「{company_style_guide}」も必ず遵守してください。
+---
+
+--- 執筆ルール ---
+1.  **提供される会話履歴（直前のセクションの内容など）と、上記「詳細なリサーチ情報」を十分に考慮し、** 前のセクションから自然につながるように、かつ、このセクション（インデックス {target_index}、見出し「{target_heading}」）の主題に沿った文章を作成してください。
+2.  **リサーチ情報で示された事実やデータに基づいて執筆し、必要に応じて、信頼できる情報源（特に公式HPなど）へのHTMLリンク (`<a href="URL">リンクテキスト</a>`) を自然な形で含めてください。** リンクテキストは具体的に、例えば会社名やサービス名、情報の内容を示すものにしてください。ただし、過剰なリンクやSEOに不自然なリンクは避けてください。リサーチ情報に記載のない情報は含めないでください。
+3.  他のセクションの内容は絶対に生成しないでください。
+4.  必ず `<p>`, `<h2>`, `<h3>`, `<ul>`, `<li>`, `<strong>`, `<em>`, `<a>` などの基本的なHTMLタグを使用し、構造化されたコンテンツを生成してください。`<h2>` タグはこのセクションの見出し「{target_heading}」にのみ使用してください。
+5.  SEOを意識し、記事全体のキーワードやこのセクションに関連するキーワードを**自然に**含めてください。（ただし、自然さを損なうような無理なキーワードの詰め込みは避けてください）
+6.  上記の【執筆スタイルとトーンについて】の指示に従い、創造性を発揮し、読者にとって価値のあるオリジナルな文章を作成してください。
+---
+
+--- **【最重要】出力形式について** ---
+あなたの応答は**必ず**、指示されたセクション（インデックス {target_index}、見出し「{target_heading}」）の**HTMLコンテンツ文字列のみ**を出力してください。
+- **JSON形式や ```html のようなマークダウン形式は絶対に使用しないでください。**
+- **「はい、以下にHTMLを記述します」のような前置きや、説明文、コメントなども一切含めないでください。**
+- **出力は `<h2...>` または `<p...>` タグから始まり、そのセクションの最後のHTMLタグで終わるようにしてください。**
+- **指定されたセクションのHTMLコンテンツだけを、そのまま出力してください。**
+"""
+        return full_prompt
+    return dynamic_instructions_func
